+++ processor/src/trace/decoder/mod.rs
@@ -59,8 +59,8 @@ fn build_aux_col_p1<E: FieldElement<BaseField = Felt>>(
         // if we skipped some cycles since the last update was processed, values in the last
         // updated row should by copied over until the current cycle.
         if result_idx < clk {
-            let last_value = result[result_idx];
-            result[(result_idx + 1)..=clk].fill(last_value);
+            let last_value = result[result_idx as usize];
+            result[(result_idx as usize + 1)..=clk as usize].fill(last_value);
         }
 
         // move the result pointer to the next row
@@ -71,7 +71,7 @@ fn build_aux_col_p1<E: FieldElement<BaseField = Felt>>(
             BlockTableUpdate::BlockStarted(_) => {
                 // when a new block is started, multiply the running product by the value
                 // representing the entry for the block in the block stack table.
-                result[result_idx] = result[clk] * row_values[started_block_idx];
+                result[result_idx as usize] = result[clk as usize] * row_values[started_block_idx];
                 started_block_idx += 1;
             }
             BlockTableUpdate::SpanExtended => {
@@ -80,7 +80,8 @@ fn build_aux_col_p1<E: FieldElement<BaseField = Felt>>(
                 // new batch.
                 let old_row_value_inv = inv_row_values[started_block_idx - 1];
                 let new_row_value = row_values[started_block_idx];
-                result[result_idx] = result[clk] * old_row_value_inv * new_row_value;
+                result[result_idx as usize] =
+                    result[clk as usize] * old_row_value_inv * new_row_value;
                 started_block_idx += 1;
             }
             BlockTableUpdate::BlockEnded(_) => {
@@ -91,19 +92,19 @@ fn build_aux_col_p1<E: FieldElement<BaseField = Felt>>(
                 let row_idx = aux_trace_hints
                     .get_block_stack_row_idx(block_id)
                     .expect("block stack row not found");
-                result[result_idx] = result[clk] * inv_row_values[row_idx];
+                result[result_idx as usize] = result[clk as usize] * inv_row_values[row_idx];
             }
             // REPEAT operation has no effect on the block stack table
-            BlockTableUpdate::LoopRepeated => result[result_idx] = result[clk],
+            BlockTableUpdate::LoopRepeated => result[result_idx as usize] = result[clk as usize],
         }
     }
 
     // at this point, block stack table must be empty - so, the last value must be ONE;
     // we also fill in all the remaining values in the column with ONE's.
-    let last_value = result[result_idx];
+    let last_value = result[result_idx as usize];
     assert_eq!(last_value, E::ONE);
-    if result_idx < result.len() - 1 {
-        result[(result_idx + 1)..].fill(E::ONE);
+    if result_idx < result.len() as u32 - 1 {
+        result[(result_idx + 1) as usize..].fill(E::ONE);
     }
 
     result
@@ -144,8 +145,8 @@ fn build_aux_col_p2<E: FieldElement<BaseField = Felt>>(
         // if we skipped some cycles since the last update was processed, values in the last
         // updated row should by copied over until the current cycle.
         if result_idx < clk {
-            let last_value = result[result_idx];
-            result[(result_idx + 1)..=clk].fill(last_value);
+            let last_value = result[result_idx as usize];
+            result[(result_idx as usize + 1)..=clk as usize].fill(last_value);
         }
 
         // move the result pointer to the next row
@@ -158,15 +159,16 @@ fn build_aux_col_p2<E: FieldElement<BaseField = Felt>>(
                 // table; in case this was a JOIN block with two children, the first child should
                 // have is_first_child set to true.
                 match *num_children {
-                    0 => result[result_idx] = result[clk],
+                    0 => result[result_idx as usize] = result[clk as usize],
                     1 => {
                         debug_assert!(!table_rows[started_block_idx].is_first_child());
-                        result[result_idx] = result[clk] * row_values[started_block_idx];
+                        result[result_idx as usize] =
+                            result[clk as usize] * row_values[started_block_idx];
                     }
                     2 => {
                         debug_assert!(table_rows[started_block_idx].is_first_child());
                         debug_assert!(!table_rows[started_block_idx + 1].is_first_child());
-                        result[result_idx] = result[clk]
+                        result[result_idx as usize] = result[clk as usize]
                             * row_values[started_block_idx]
                             * row_values[started_block_idx + 1];
                     }
@@ -188,7 +190,7 @@ fn build_aux_col_p2<E: FieldElement<BaseField = Felt>>(
                 let row_idx = aux_trace_hints
                     .get_block_hash_row_idx(parent_id, false)
                     .expect("block hash row not found");
-                result[result_idx] = result[clk] * row_values[row_idx];
+                result[result_idx as usize] = result[clk as usize] * row_values[row_idx];
             }
             BlockTableUpdate::BlockEnded(is_first_child) => {
                 // when END operation is executed, we need to remove an entry for the block from
@@ -200,19 +202,19 @@ fn build_aux_col_p2<E: FieldElement<BaseField = Felt>>(
                 let row_idx = aux_trace_hints
                     .get_block_hash_row_idx(parent_id, *is_first_child)
                     .expect("block hash row not found");
-                result[result_idx] = result[clk] * inv_row_values[row_idx];
+                result[result_idx as usize] = result[clk as usize] * inv_row_values[row_idx];
             }
             // RESPAN operation has no effect on the block hash table
-            BlockTableUpdate::SpanExtended => result[result_idx] = result[clk],
+            BlockTableUpdate::SpanExtended => result[result_idx as usize] = result[clk as usize],
         }
     }
 
     // at this point, block hash table must be empty - so, the last value must be ONE;
     // we also fill in all the remaining values in the column with ONE's.
-    let last_value = result[result_idx];
+    let last_value = result[result_idx as usize];
     assert_eq!(last_value, E::ONE);
-    if result_idx < result.len() - 1 {
-        result[(result_idx + 1)..].fill(E::ONE);
+    if result_idx < result.len() as u32 - 1 {
+        result[(result_idx as usize + 1)..].fill(E::ONE);
     }
 
     result
@@ -249,13 +251,13 @@ fn build_aux_col_p3<E: FieldElement<BaseField = Felt>>(
 
         // if we skipped some cycles since the last update was processed, values in the last
         // updated row should by copied over until the current cycle.
-        if result_idx < clk {
+        if result_idx < clk as usize {
             let last_value = result[result_idx];
-            result[(result_idx + 1)..=clk].fill(last_value);
+            result[(result_idx + 1)..=clk as usize].fill(last_value);
         }
 
         // apply the relevant updates to the column
-        result_idx = clk + 1;
+        result_idx = clk as usize + 1;
         match update {
             OpGroupTableUpdate::InsertRows(num_op_groups) => {
                 // if the rows were added, multiply the current value in the column by the values
@@ -264,7 +266,7 @@ fn build_aux_col_p3<E: FieldElement<BaseField = Felt>>(
                 for i in 1..(*num_op_groups as usize) {
                     value *= row_values[inserted_group_idx + i];
                 }
-                result[result_idx] = result[clk] * value;
+                result[result_idx] = result[clk as usize] * value;
 
                 // advance the inserted group pointer by the number of inserted rows
                 inserted_group_idx += *num_op_groups as usize;
@@ -272,7 +274,7 @@ fn build_aux_col_p3<E: FieldElement<BaseField = Felt>>(
             OpGroupTableUpdate::RemoveRow => {
                 // if a row was removed, divide the current value in the column by the value
                 // of the row
-                result[result_idx] = result[clk] * inv_row_values[removed_group_idx];
+                result[result_idx] = result[clk as usize] * inv_row_values[removed_group_idx];
 
                 // advance the removed group pointer by one
                 removed_group_idx += 1;
@@ -295,6 +297,6 @@ fn build_aux_col_p3<E: FieldElement<BaseField = Felt>>(
 // ================================================================================================
 
 /// Returns the value in the block address column at the specified row.
-fn get_block_addr(main_trace: &Matrix<Felt>, row_idx: usize) -> Felt {
-    main_trace.get(ADDR_COL_IDX, row_idx)
+fn get_block_addr(main_trace: &Matrix<Felt>, row_idx: u32) -> Felt {
+    main_trace.get(ADDR_COL_IDX, row_idx as usize)
 }
